create and replace new component
