+++ src/Componets/App.jsx
@@ -1,34 +1,12 @@
-import React, { useReducer } from 'react';
-import UseReduserCom from './reactHooks/UseReducer/UseReduserCom';
-import UseReduserCom2 from './reactHooks/UseReducer/UseReduserCom2';
-import UseReduserCom3 from './reactHooks/UseReducer/UseReduserCom3';
-import ComponentA from './reactHooks/useRandUseC/ComponentA';
-import { CountState } from '../Data/context';
+import React from 'react';
+import DataFetching from '../Componets/reactHooks/useReducerFetch/DateFetching';
 
-const reducer = (state, action) => {
-  switch (action) {
-    case 'increment':
-      return state + 1;
-    case 'decrement':
-      return state - 1;
-    case 'restart':
-      return 0;
-    default:
-      return state;
-  }
-};
-
-function App() {
-  const [state, dispatch] = useReducer(reducer, 0);
+const App = () => {
   return (
-    <CountState.Provider value={{ count: state, countDispatch: dispatch }}>
-      Count = {state}
-      {/* <UseReduserCom /> */}
-      {/* <UseReduserCom2 /> */}
-      {/* <UseReduserCom3 /> */}
-      <ComponentA />
-    </CountState.Provider>
+    <div>
+      <DataFetching />
+    </div>
   );
-}
+};
 
 export default App;
