+++ src/Componets/reactHooks/useReducerFetch/DateFetching.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const DateFetching = () => {
+  return <div>data</div>;
+};
+
+export default DateFetching;
